move hash final worker into new file
